// generated/src/app/api-models/sec-user-param-profile.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SecUserParamProfile {
  userId: string;
  secParamSeq: number;
  columnName: string;
  columnOccurrence: number;
  columnType: string;
  operator: string;
  fromValue: string;
  thruValue: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}